Move certificate template logic into its own component
